Add Dockerfile and docker-compose.yml
These files build and run pulumigraph without requiring you to have
dotnet or neo4j on your machine.

Currently this uses the pulumigraph tool released to nuget. That version
connects to neo4j on "localhost".  In this stack, neo4j is running in a
separate container "neo4j" on the same Docker netowrk. So either the
`docker-compose build` step has to build pulumigraph from source (I've
changed the connection string to "neo4j"), or the tool needs to be able
to take a target host name or connection string as a parameter.